PhishingDetectionSystem/backend/requirements.txt
flask
flask-cors
dnspython
transformers
torch
scipy
numpy
waitress
